Refactor and bug fixes

Approved-by: Edward Campos
Approved-by: Osama Bin Waseem

+++ project/static/js/patient-account.js
@@ -5,39 +5,18 @@ function deleteClinician(email) {
         return
     }
 
+    const confirmation = confirm("Are you sure you would like to remove this clinician from your account?");
+    if (confirmation) {
     fetch("/patient/account/" + email, { method: "DELETE", })
-    .then(res => {
-    if (!res.ok) {
-        console.log("Error", res);
-    } else {
-        console.log("Success", res);
-        location.reload();
-    }
-    });
-}
-
-clinicians = document.currentScript.getAttribute('clinicians').split(",");
-window.onload = function() {
-    var ul = document.getElementById('cl-list-container');
-
-    if (clinicians == null || clinicians == '' || clinicians.length == 0) {
-        empty = document.createElement("div");
-        empty.appendChild(document.createTextNode("You have not linked any clinicians."))
-        ul.appendChild(empty);
-    }
-    else {
-        for (var i = 0; i < clinicians.length; i++) {
-            var li = document.createElement("li");
-            li.appendChild(document.createTextNode(clinicians[i]));
-
-            var span = document.createElement("span");
-            span.setAttribute('class', 'close');
-            span.setAttribute('id', clinicians[i]);
-            span.setAttribute('onclick', 'deleteClinician(id)');
-            span.appendChild(document.createTextNode("x"));
-
-            li.appendChild(span);
-            ul.appendChild(li);
+        .then(res => {
+        if (!res.ok) {
+            console.log("Error", res);
+        } else {
+            console.log("Success", res);
+            location.reload();
         }
-    }
-};
+    });
+  }
+
+    
+}
